+++ distribution/src/bin/elasticsearch-env.bat
@@ -15,7 +15,7 @@ for %%I in ("%ES_HOME%..") do set ES_HOME=%%~dpfI
 
 rem now set the classpath
 set ES_CLASSPATH=!ES_HOME!\lib\*
-set LAUNCHERS_CLASSPATH=!ES_CLASSPATH!;!ES_HOME!\lib\launchers\*;!ES_HOME!\lib\java-version-checker\*
+set SERVER_CLI_CLASSPATH=!ES_CLASSPATH!;!ES_HOME!\lib\tools\server-cli\*
 
 set HOSTNAME=%COMPUTERNAME%
