+++ distribution/tools/server-cli/build.gradle
@@ -1,0 +1,34 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0 and the Server Side Public License, v 1; you may not use this file except
+ * in compliance with, at your election, the Elastic License 2.0 or the Server
+ * Side Public License, v 1.
+ */
+import de.thetaphi.forbiddenapis.gradle.CheckForbiddenApis
+
+apply plugin: 'elasticsearch.build'
+
+dependencies {
+  implementation "org.yaml:snakeyaml:${versions.snakeyaml}"
+  testImplementation "com.carrotsearch.randomizedtesting:randomizedtesting-runner:${versions.randomizedrunner}"
+  testImplementation "junit:junit:${versions.junit}"
+  testImplementation "org.hamcrest:hamcrest:${versions.hamcrest}"
+}
+
+tasks.withType(CheckForbiddenApis).configureEach {
+  replaceSignatureFiles 'jdk-signatures', 'snakeyaml-signatures'
+}
+
+tasks.named("testingConventions").configure {
+  naming.clear()
+  naming {
+    Tests {
+      baseClass 'org.elasticsearch.server.cli.LaunchersTestCase'
+    }
+  }
+}
+
+["javadoc", "loggerUsageCheck", "jarHell"].each { tsk ->
+  tasks.named(tsk).configure { enabled = false }
+}
