+++ distribution/tools/server-cli/licenses/snakeyaml-LICENSE.txt
