+++ distribution/tools/server-cli/src/test/resources/config/elasticsearch.yml
